# jeekite-online/src/main/java/com/jeekite/online/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "微服务工程代码生成器,生成Dao、Service、Controller、Pojo、Feign、Swagger的JSON脚本",
    "version": "1.0.0",
    "title": "dudianlong",
    "termsOfService": " https://github.com/dwx617676/jeekite-pro.git",
    "contact": {
      "email": "13164687433@163.COM"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "petstore.swagger.io",
  "basePath": "/v2",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/sysDept/{id}": {
      "delete": {
        "tags": [
          "SysDeptController"
        ],
        "summary": "根据ID删除SysDept",
        "description": "根据ID删除SysDept方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除SysDept方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除SysDept"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "SysDeptController"
        ],
        "summary": "根据ID修改SysDept",
        "description": "根据ID修改SysDept方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改SysDept方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "SysDept",
            "description": "传入SysDept的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SysDept"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改SysDept"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "SysDeptController"
        ],
        "summary": "根据ID查询SysDept",
        "description": "根据ID查询SysDept方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改SysDept方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改SysDept"
            ,
            "schema": {
              "$ref": "#/definitions/Result«SysDept»"
            }
          }
        }
      }
    },
    "/sysDept/search": {
      "post": {
        "tags": [
          "SysDeptController"
        ],
        "summary": "不带分页条件搜索SysDept",
        "description": "不带分页条件搜索SysDept方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "SysDept",
            "description": "传入SysDept的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SysDept"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索SysDept"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SysDept»»"
            }
          }
        }
      }
    },
    "/sysDept/search/{page}/{size}": {
      "get": {
        "tags": [
          "SysDeptController"
        ],
        "summary": "分页列表查询SysDept",
        "description": "分页列表查询SysDept方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索SysDept"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SysDept»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "SysDeptController"
        ],
        "summary": "分页条件搜索SysDept",
        "description": "分页条件搜索SysDept方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "SysDept",
            "description": "传入SysDept的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SysDept"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索SysDept"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SysDept»»"
            }
          }
        }
      }
    },
...
    "/sysDict/{id}": {
      "delete": {
        "tags": [
          "SysDictController"
        ],
        "summary": "根据ID删除SysDict",
        "description": "根据ID删除SysDict方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除SysDict方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除SysDict"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "SysDictController"
        ],
        "summary": "根据ID修改SysDict",
        "description": "根据ID修改SysDict方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改SysDict方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "SysDict",
            "description": "传入SysDict的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SysDict"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改SysDict"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "SysDictController"
        ],
        "summary": "根据ID查询SysDict",
        "description": "根据ID查询SysDict方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改SysDict方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改SysDict"
            ,
            "schema": {
              "$ref": "#/definitions/Result«SysDict»"
            }
          }
        }
      }
    },
    "/sysDict/search": {
      "post": {
        "tags": [
          "SysDictController"
        ],
        "summary": "不带分页条件搜索SysDict",
        "description": "不带分页条件搜索SysDict方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "SysDict",
            "description": "传入SysDict的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SysDict"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索SysDict"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SysDict»»"
            }
          }
        }
      }
    },
    "/sysDict/search/{page}/{size}": {
      "get": {
        "tags": [
          "SysDictController"
        ],
        "summary": "分页列表查询SysDict",
        "description": "分页列表查询SysDict方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索SysDict"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SysDict»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "SysDictController"
        ],
        "summary": "分页条件搜索SysDict",
        "description": "分页条件搜索SysDict方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "SysDict",
            "description": "传入SysDict的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SysDict"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索SysDict"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SysDict»»"
            }
          }
        }
      }
    },
...
    "/sysDictDetail/{id}": {
      "delete": {
        "tags": [
          "SysDictDetailController"
        ],
        "summary": "根据ID删除SysDictDetail",
        "description": "根据ID删除SysDictDetail方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除SysDictDetail方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除SysDictDetail"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "SysDictDetailController"
        ],
        "summary": "根据ID修改SysDictDetail",
        "description": "根据ID修改SysDictDetail方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改SysDictDetail方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "SysDictDetail",
            "description": "传入SysDictDetail的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SysDictDetail"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改SysDictDetail"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "SysDictDetailController"
        ],
        "summary": "根据ID查询SysDictDetail",
        "description": "根据ID查询SysDictDetail方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改SysDictDetail方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改SysDictDetail"
            ,
            "schema": {
              "$ref": "#/definitions/Result«SysDictDetail»"
            }
          }
        }
      }
    },
    "/sysDictDetail/search": {
      "post": {
        "tags": [
          "SysDictDetailController"
        ],
        "summary": "不带分页条件搜索SysDictDetail",
        "description": "不带分页条件搜索SysDictDetail方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "SysDictDetail",
            "description": "传入SysDictDetail的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SysDictDetail"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索SysDictDetail"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SysDictDetail»»"
            }
          }
        }
      }
    },
    "/sysDictDetail/search/{page}/{size}": {
      "get": {
        "tags": [
          "SysDictDetailController"
        ],
        "summary": "分页列表查询SysDictDetail",
        "description": "分页列表查询SysDictDetail方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索SysDictDetail"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SysDictDetail»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "SysDictDetailController"
        ],
        "summary": "分页条件搜索SysDictDetail",
        "description": "分页条件搜索SysDictDetail方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "SysDictDetail",
            "description": "传入SysDictDetail的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SysDictDetail"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索SysDictDetail"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SysDictDetail»»"
            }
          }
        }
      }
    },
...
    "/sysGenerator/search": {
      "post": {
        "tags": [
          "SysGeneratorController"
        ],
        "summary": "不带分页条件搜索SysGenerator",
        "description": "不带分页条件搜索SysGenerator方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "SysGenerator",
            "description": "传入SysGenerator的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SysGenerator"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索SysGenerator"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SysGenerator»»"
            }
          }
        }
      }
    },
...
    "/sysJob/search": {
      "post": {
        "tags": [
          "SysJobController"
        ],
        "summary": "不带分页条件搜索SysJob",
        "description": "不带分页条件搜索SysJob方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "SysJob",
            "description": "传入SysJob的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/SysJob"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索SysJob"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«SysJob»»"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Result": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "type": "object",
          "description": "逻辑数据"
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
...
    "Result«SysDept»": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "description": "逻辑数据",
          "$ref": "#/definitions/SysDept"
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
...
  },
...
}